Created the data and binary files. Programmed the scroller
Due to a PC problem last Tuesday, and complete loss of data. I decided to start this project from scratch, and managed to catch up from last time - In a day.  The scroller is now in place, and the player can move around and shoot bullets. Laser gates can also activate/deactivate.