Collision detection added and also new hud
The game now has sprite/background and sprite/sprite collision. When the player is hit by a deadly background or alien. It should lose a shield. When all shields are lost - game over (Not implemented yet).

Updated the HUD to make one of the panels the correct size by expanding the shield icon.